move scatter_data to spawn
